rss module: prettify & reorganize to allow for easily adding extra modules
